Rename XInput patch to Gamepad
